+++ lib/secondary_screens/professionals_details_screen/service_tile.dart
@@ -15,7 +15,7 @@ class ServiceTiles extends StatefulWidget {
   final String professionalId;
   final String serviceTitle;
   final String serviceDescription;
-  final int servicePrice;
+  final double servicePrice;
 
   ServiceTiles({
     @required this.serviceId,
